Made cleanup task only run on OpenShift

+++ post-install.sh
@@ -18,21 +18,27 @@ $bower install
 echo 'running gulp...'
 $gulp
 
-echo 'deployment build complete, starting cleanup...'
-echo 'removing bower_components...'
-rm -r ./bower_components
-echo 'uninstalling bower...'
-npm uninstall bower
-echo 'removing bower.json...'
-rm ./bower.json
-echo 'uninstalling gulp...'
-npm uninstall gulp
-echo 'removing gulpfile.js'
-rm ./gulpfile.js
-echo 'removing sass source...'
-rm -r ./css
-echo 'removing deployment script...'
-rm ./deploy
-echo 'cleanup complete'
+echo -n 'deployment build complete,'
+
+if [ -z $OPENSHIFT_REPO_DIR ]; then
+    echo ' skipping cleanup.'
+else
+    echo ' starting cleanup...'
+    echo 'removing bower_components...'
+    rm -r ./bower_components
+    echo 'uninstalling bower...'
+    npm uninstall bower
+    echo 'removing bower.json...'
+    rm ./bower.json
+    echo 'uninstalling gulp...'
+    npm uninstall gulp
+    echo 'removing gulpfile.js'
+    rm ./gulpfile.js
+    echo 'removing sass source...'
+    rm -r ./css
+    echo 'removing deployment script...'
+    rm ./deploy
+    echo 'cleanup complete'
+fi
 
 echo 'done'
